Part B.2 - Web App/requirements.txt
Flask
faiss-cpu
numpy
sentence-transformers   